requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
pyzbar==0.1.9
opencv-python==4.8.1.78
requests==2.31.0
Flask-WTF==1.1.1
Flask-Caching==2.0.2
redis==5.0.1
celery==5.3.4
Flask-Talisman==1.1.0